HTTP Block, and File Input Block : Move common metadata functions to shared utility file avoiding duplication, adds raw byte count, documentation.
